Added code quality annotations "Exact" (as vanilla), "Compatible" (equal in terms of result or has cfg option to be), "Suspicious" (differs from original by result and implementation)
Commented out killough's code on thinkers. Will all the respect,
it was unused, seemed to be only an unsuccessfull attempt to
include in Mocha Doom, and only took extra memory and cycles to
waste. Could possibly implement something new in future
(have profeciency in data structures, even a Trie for thinkers
is not out of question at the moment)